+++ far/cmdline.hpp
@@ -71,7 +71,7 @@ public:
 
 	virtual int ProcessKey(const Manager::Key& Key) override;
 	virtual int ProcessMouse(const MOUSE_EVENT_RECORD *MouseEvent) override;
-	virtual __int64 VMProcess(int OpCode,void *vParam=nullptr,__int64 iParam=0) override;
+	virtual long long VMProcess(int OpCode, void* vParam = nullptr, long long iParam=0) override;
 
 	const string& GetCurDir() const { return m_CurDir; }
 	void SetCurDir(const string& CurDir);
